Small fixes to complile and test with CUDA 6.5, GCC 4.9.2 and Ubuntu 14.04.

+++ GDelaunay/Common/CudaWrapper.h
@@ -144,7 +144,7 @@ inline void cuPrintMemory( const char* inStr )
     size_t total;
     CudaSafeCall( cudaMemGetInfo( &free, &total ) );
 
-    printf( "[%s] Memory used: %d MB\n", inStr, ( total - free ) / MegaByte );
+    printf( "[%s] Memory used: %d MB\n", inStr, (int) ( total - free ) / MegaByte );
 
     return;
 }
